pyproject.toml
[tool.poetry]
name = "floshiptest"
version = "0.1.0"
description = ""
authors = ["andoleg <and.olega@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.6"
psycopg2-binary = "^2.8.6"
drf-yasg = "^1.20.0"
djangorestframework = "^3.12.2"
django-choices = "^1.7.1"
celery = {extras = ["rabbitmq"], version = "^5.0.5"}
django-3-jet = "^1.0.8"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
